Jobs display based on Location and Company
When you select Existing Jobs, you will pick a company. That in turn
will populate a drop down list with all locations for that company. Upon
selecting a location and clicking submit, a table populates with all
jobs for that location. No buttons implemented yet to Update a job yet